fix(moe-q4k-mma): F32 GEMM intermediate (was F16 — drift compounded to NaN)
Smoke test against the production dp4a kernel now shows max rel err
0.014% (was 0.26% with F16 intermediate). F16 rounding step in the
GEMM output compounded through 30 cascading layers to NaN logits on
real gemma4:26b prefill.

Added moe_batched_gelu_mul_scatter_f32_to_f32 kernel + FFI.
Orchestrator allocates F32 gemm_alloc and dispatches the F32 scatter.

**Bench result on real gemma4:26b prefill (LLMSERVER_MOE_MMA=1):**
- llmserver dp4a: 2172 prompt-tok/s
- llmserver MMA:   838 prompt-tok/s    (-61% vs dp4a)
- Ollama:        26982 prompt-tok/s
- decode: 102.9 vs Ollama 106.1 = -3%  (still TIED)

Model output now coherent (no NaN). But the MMA path is SLOWER than
dp4a in practice — same lesson as the 5 prior wiring variants:
per-expert dispatch with multi-kernel dispatch (gather+MMA+scatter)
can't beat dp4a's fused per-(token,expert) kernel for this regime.
The structural fix for the -91.8% gemma4:26b prefill gap needs to
GO BEYOND just kernel choice (e.g., kernel-level fusion of all three
phases into one launch, persistent CTA tiling, or a Q4_K-direct MMQ
kernel modeled on llama.cpp's mmq.cu — multi-week).

+++ candle-kernels/src/moe/moe_q4k_mma_batched.cu
@@ -151,7 +151,7 @@ extern "C" __global__ void moe_q4k_mma_batched_gate_up_kernel(
     const void * __restrict__ inputs_y,          // [N_active, max_n_e, K/32] Q8_1
     const int32_t * __restrict__ active_expert_ids,
     const int32_t * __restrict__ expert_offsets,
-    void * __restrict__ dst,                     // [N_active, max_n_e, 2N] F16
+    void * __restrict__ dst,                     // [N_active, max_n_e, 2N] F32
     int num_experts,
     int max_n_e,
     int two_n,
@@ -306,19 +306,19 @@ extern "C" __global__ void moe_q4k_mma_batched_gate_up_kernel(
     //   D3 → dst[act_idx, n_base+2tj+1, row_b]
     const int in_row_0 = n_base + 2 * tj + 0;
     const int in_row_1 = n_base + 2 * tj + 1;
-    __half * dst_base = (__half *) dst + (size_t)act_idx * max_n_e * two_n;
+    float * dst_base = (float *) dst + (size_t)act_idx * max_n_e * two_n;
 
-if (va && in_row_0 < n_e) {
-        dst_base[(size_t)in_row_0 * two_n + row_a] = __float2half(acc_d_0);
+    if (va && in_row_0 < n_e) {
+        dst_base[(size_t)in_row_0 * two_n + row_a] = acc_d_0;
     }
     if (va && in_row_1 < n_e) {
-        dst_base[(size_t)in_row_1 * two_n + row_a] = __float2half(acc_d_1);
+        dst_base[(size_t)in_row_1 * two_n + row_a] = acc_d_1;
     }
     if (vb && in_row_0 < n_e) {
-        dst_base[(size_t)in_row_0 * two_n + row_b] = __float2half(acc_d_2);
+        dst_base[(size_t)in_row_0 * two_n + row_b] = acc_d_2;
     }
     if (vb && in_row_1 < n_e) {
-        dst_base[(size_t)in_row_1 * two_n + row_b] = __float2half(acc_d_3);
+        dst_base[(size_t)in_row_1 * two_n + row_b] = acc_d_3;
     }
 
     (void)num_experts;
